Attempt to transition to Neovim, again
I just wasn't using it while it was there without the alias. So here
goes. The :terminal experience is still going to take a lot of getting
used to. It's no longer as easy to work with Vim-test, can't navigate
longer output (which is what you get when things go wrong, right?), old
output completely disappears unless you run the specs again ("what was
the line number again?"). Urgh, I'll try and work it out.